// packages/indexer/src/models/activities/activities-entity.ts
export enum ActivityType {
  sale = "sale",
  ask = "ask",
  transfer = "transfer",
  mint = "mint",
  bid = "bid",
  bid_cancel = "bid_cancel",
  ask_cancel = "ask_cancel",
}

// Define the fields required to create a new activity
export type ActivitiesEntityInsertParams = {
  type: ActivityType;
  hash: string;
  contract: string;
  collectionId: string;
  tokenId: string | null;
  orderId: string | null;
  fromAddress: string;
  toAddress: string | null;
  price: number;
  amount: number;
  blockHash: string | null;
  eventTimestamp: number;
  metadata?: ActivityMetadata;
};
// ...
export type ActivityMetadata = {
  transactionHash?: string;
  logIndex?: number;
  batchIndex?: number;
  orderId?: string;
  orderSide?: string;
  orderSourceIdInt?: number;
  orderKind?: string;
  orderMetadata?: Record<string, unknown>;
  orderCriteria?: Record<string, unknown>;
};